feat: add `wrap_exceptions` option to exception handler. (#363)
There's a new option to make the exception handler more configurable.  Here's the way it'll work going forward:

If you don't want any error message mapping and you don't want the exception handler to wrap the exception in a RepositoryError, then you can set `wrap_exceptions` to `False` on the repo.

If you want the wrapped repository error, but have it skip the "better error messages" bit, send in `None` for the error messages here.  We differentiate Empty from None, so that we can disable the message section when the default is not Empty but is None.

When left alone, it's the current behavior